NCI-Agency/anet#391 - refactored email worker.
Extracted database operations for emails into a DAO, and made inner classes in
AnetEmailWorker top-level classes.
